requirements.txt
flask>=3.0.0
python-telegram-bot>=21.5
anthropic>=0.40.0
google-auth>=2.35.0
google-auth-oauthlib>=1.2.1
google-api-python-client>=2.150.0
aiohttp>=3.10.0
python-dotenv>=1.0.0
requests>=2.32.0
openai>=1.0.0
